Add lane detectors.
In this version, we add a basic lane detector. Since all debug outputs
are not disabled, you can see in the output graph that, the car is in
the middle part of the road, the left lane is marked with green, and the
right lane is marked with red.

A simple way to use it is by connecting all those points, draw a
polygon, and calculate the uncovered area from a user uploaded one.